Updated doc + started code
- Started working on code for Arduino
- Added doc for ampli-op, capacitors
