+++ Documentation/LM1085IS-12NOPB - LDO, FIXED, 12V, 3A, TO-263-3.pdf
