+++ Documentation/LVR025K - POLYSWITCH, 0.25A.pdf
