+++ Documentation/T491C106K035AT - CAP, TANT, 10UF, 35V, CASE C.pdf
